Partially implemented (ported) the GP "TEE Internal API"

+++ tee_internal/include/utee_types.h
@@ -56,17 +56,17 @@ enum utee_cache_operation {
 };
 
 struct utee_params {
-	uint64_t types;
+	uint32_t types;
 	/* vals[n * 2]	   corresponds to either value.a or memref.buffer
 	 * vals[n * 2 + ]  corresponds to either value.b or memref.size
 	 * when converting to/from struct tee_ta_param
 	 */
-	uint64_t vals[TEE_NUM_PARAMS * 2];
+	uint32_t vals[TEE_NUM_PARAMS * 2];
 };
 
 struct utee_attribute {
-	uint64_t a;	/* also serves as a pointer for references */
-	uint64_t b;	/* also serves as a length for references */
+	uint32_t a;	/* also serves as a pointer for references */
+	uint32_t b;	/* also serves as a length for references */
 	uint32_t attribute_id;
 };
